Add params logging on query failure
This feels a bit dumb, but the filter behavior has been inconsistent,
and I'd like to be able to see what the params were if/when the failure
occurs. Can use this to reproduce, as I've been unable to do so (at all)
when running ibutsu locally with a smaller data set.