Replace LogViewer with CodeBlock and Expandable
Move the download button to the top and make the card content
expandable.

This provides for collapsing long logs, or screenshots when necessary.

Also makes the download button more accessible.

I tried to debug the LogViewer, but it was very consistently only
rendering the data after some time. Logging and everything else
attempted confirmed that the blob was fetched, it's type and content was
valid and in state - but the component would not render the text. The
behaviors reported for non-rendering without search, and issues with
scrolling once rendered, persisted. I made the string an array, I tried
with/without wrapping and line numbers enabled. Tried using a ref to
force rendering via key prop, tried forcing height, setting item count,
forcing scroll to 0 line on render.

Couldn't locate any upstream issues specifically for this, but did
locate a related issue for when the LogViewer is in an
ExpandableSection:
https://github.com/patternfly/react-log-viewer/issues/19#issue-1626199032

I'm trying to stick with Patternfly components here, there are plenty of
great upstream tools available. CodeBlock is functional but not
attractive for this use case